Add support for passive connection lost checking
When both ends of a WebSocket connection need to know when the
connection is lost, both sending pings is redundant, as a received ping
already indicates that the peer is still connected.

Add an optional boolean parameter to the setConnectionLostTimeout()
method, which when set to true makes the connection lost cheking
passive, i.e. the check never sends out pings, but only checks whether
an incoming ping was received recently enough.

For passive connection lost checking, the peer must be known to do
active connection lost checking with a timeout equal (or lower) than the
one locally used.