Fix null-pointer vulnerability and restore Android compatibility
The loop checking the list of input extensions passed as a constructor
parameter was blindly calling getClass() on any element in the list.
This is not null safe and broke compatibility with some Android
versions.

Change the code to a null-safe instanceof check, which also restores
Android compatibility.

Fixes Issue #1353.