Fix Per-Message Compression Extension support
The per-message compression extension was badly broken, leading to
exceptions and loss of communication between client and server.

Fix the extension negotiation to correctly send the current state of the
client_no_context_takeover and server_no_context_takeover configuration
variables, and ORing the received extensions with the received ones, to
ensure client and server end up with the same configuration.

Fix the application of these configuration variables, which must be
swapped on the client vs. the server: The configuration of the deflater
on the client side must affect the inflater on the server side, and vice
versa.

Remove the requestedParameters map as there is no need for it.

This fixes issue #1496